refactor: ErrorBoundary, ContentBox 컴포넌트 common 디렉터리로 이동
