+++ src/component/common/ContentBox/style.js
